Relocate cypress cloud package
